Subtitle renderer: Use short integers for the Gaussian blur filter kernel.
No output change, will be useful for future work.

+++ src/Subtitles/SeparableFilter.h
@@ -27,7 +27,8 @@
 // Filter an image in horizontal direction with a one-dimensional filter
 // PixelWidth is the distance in bytes between pixels
 template<ptrdiff_t PixelDist>
-void SeparableFilterX(unsigned char* src, unsigned char* dst, int width, int height, ptrdiff_t stride, int* kernel, int kernel_size, int divisor)
+void SeparableFilterX(unsigned char* src, unsigned char* dst, int width, int height, ptrdiff_t stride,
+                      short* kernel, int kernel_size, int divisor)
 {
     int* tmp = DEBUG_NEW int[width];
 
@@ -68,7 +69,8 @@ void SeparableFilterX(unsigned char* src, unsigned char* dst, int width, int hei
 // Filter an image in vertical direction with a one-dimensional filter
 // PixelWidth is the distance in bytes between pixels
 template<ptrdiff_t PixelDist>
-void SeparableFilterY(unsigned char* src, unsigned char* dst, int width, int height, ptrdiff_t stride, int* kernel, int kernel_size, int divisor)
+void SeparableFilterY(unsigned char* src, unsigned char* dst, int width, int height, ptrdiff_t stride,
+                      short* kernel, int kernel_size, int divisor)
 {
     int* tmp = DEBUG_NEW int[width];
 
@@ -108,10 +110,10 @@ void SeparableFilterY(unsigned char* src, unsigned char* dst, int width, int hei
 
 static inline double NormalDist(double sigma, double x)
 {
-    if (sigma <= 0 && x == 0) {
-        return 1;
-    } else if (sigma <= 0) {
-        return 0;
+    if (sigma <= 0.0 && x == 0.0) {
+        return 1.0;
+    } else if (sigma <= 0.0) {
+        return 0.0;
     } else {
         return exp(-(x * x) / (2 * sigma * sigma)) / (sigma * sqrt(2 * M_PI));
     }
@@ -119,24 +121,26 @@ static inline double NormalDist(double sigma, double x)
 
 
 struct GaussianKernel {
-    int* kernel;
+    short* kernel;
     int width;
     int divisor;
+
     inline GaussianKernel(double sigma) {
-        width = (int)(sigma * 3 + 0.5) | 1; // binary-or with 1 to make sure the number is odd
+        width = (int)(sigma * 3.0 + 0.5) | 1; // binary-or with 1 to make sure the number is odd
         if (width < 3) {
             width = 3;
         }
-        kernel = DEBUG_NEW int[width];
-        kernel[width / 2] = (int)(NormalDist(sigma, 0) * 255);
+        kernel = DEBUG_NEW short[width];
+        kernel[width / 2] = (short)(NormalDist(sigma, 0.0) * 255);
         divisor = kernel[width / 2];
         for (int x = width / 2 - 1; x >= 0; x--) {
-            int val = (int)(NormalDist(sigma, width / 2 - x) * 255 + 0.5);
+            short val = (short)(NormalDist(sigma, width / 2 - x) * 255 + 0.5);
             divisor += val * 2;
             kernel[x] = val;
             kernel[width - x - 1] = val;
         }
     }
+
     inline ~GaussianKernel() {
         delete [] kernel;
     }
